back/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==13.0.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
timedelta==2020.12.3
